+++ src/comments_check.c
@@ -4,7 +4,7 @@
 
 int check_single_comment(char* str)
 {
-    unsigned int len = strlen(str);
+    size_t len = strlen(str);
     for (unsigned int i = 0; i < len; i++) {
         if (str[i] == '/' && str[i + 1] == '/') {
             return 1;
@@ -14,7 +14,7 @@ int check_single_comment(char* str)
 }
 int check_multiline_comment_begin(char* str)
 {
-    unsigned int len = strlen(str);
+    size_t len = strlen(str);
     for (unsigned int i = 0; i < len; i++) {
         if (str[i] == '/' && str[i + 1] == '*') {
             return 1;
@@ -24,7 +24,7 @@ int check_multiline_comment_begin(char* str)
 }
 int check_multiline_comment_end(char* str)
 {
-    unsigned int len = strlen(str);
+    size_t len = strlen(str);
     for (unsigned int i = 0; i < len; i++) {
         if (str[i] == '*' && str[i + 1] == '/') {
             return 1;
